# Use global string map for MatchType.String()