@colega 2021-08-23
Updated to use an array instead of a map, which is even better at 1ns/op without any extra compromises:

```
name \ time/op    old          global_map  stringer    array
MatchType_String  53.6ns ± 1%  4.1ns ± 1%  2.8ns ± 1%  1.0ns ± 1%

name \ alloc/op   old          global_map  stringer    array
MatchType_String   0.00B       0.00B       0.00B       0.00B

name \ allocs/op  old          global_map  stringer    array
MatchType_String    0.00        0.00        0.00        0.00
```